requirements.txt
celery
fastapi
uvicorn
python-dotenv
redis
flower

# dev
pytest
requests